extract the code related to mocking the DB NOW() function out from testhelpers.TextContext into a new package databasetest in order to use it in unit tests
